# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Affilate API",
        "version": "1.0.0",
        "description": "API documentation for the Affilate project"
    },
    "servers": [
        {
            "url": "http://localhost:3001",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/products/{id}": {
            "get": {
                "summary": "Get a product by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    }
}